Cambios Larry Falta punto 7 auto

+++ Parte2/Parte2b/p2_815877-Perez_Salazar_841983-Larraz_Jordan/gestorAlarma.h
@@ -10,7 +10,7 @@ typedef struct {
 		uint8_t enUso;
 		uint32_t retardo;
 		uint32_t auxData;
-		uint32_t esPeriodica;
+		uint8_t esPeriodica;
 		uint32_t contador;
 } Alarma;
